Iniaitl timelapsed and preview function working

+++ html/index.php
@@ -13,35 +13,79 @@
 		return $size;
 	}
 
+
+	function sendCommand($command)
+	{
+		$service_port = "1337";
+		$address = "127.0.0.1";
+
+		$socket = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
+		if ($socket === false) {
+			    echo "socket_create() failed: reason: " . socket_strerror(socket_last_error()) . "\n";
+		}
+
+		$result = socket_connect($socket, $address, $service_port);
+		if ($result === false) {
+		    echo "socket_connect() failed.\nReason: ($result) " . socket_strerror(socket_last_error($socket)) . "\n";
+		} 
+
+		$in = $command."\r\n";
+		socket_write($socket, $in, strlen($in));
+
+		$in = "QUIT\r\n";
+		socket_write($socket, $in, strlen($in));
+
+		while ($out = socket_read($socket, 2048)) {
+			// Waiting for connection to be terminated
+		}
+
+		socket_close($socket);
+	}
+
 	$currentstatus = $status = yaml_parse(file_get_contents("status.dat"));
 	
 	$stpl = new Mikjaer\SimpleTpl\SimpleTpl();
 
 	$validPages = array("camera","files","settings","wifi","bluetooth");
 	
+
+	if (isset($_REQUEST["ajax"]))
+	{
+		switch ($_REQUEST["ajax"])
+		{
+			case "preview":
+				sendCommand("PREVIEW");
+				print json_encode("/preview.jpg?".time());
+				die();
+		}
+	}
+
 	if (!in_array($_REQUEST["p"],$validPages))
 		header("location: ?p=camera");
 	$stpl->assign("currentPage", $currentPage=$_REQUEST["p"]);
 
-
 	if ($currentPage == "camera")
 	{
-		if ($_REQUEST["a"] == "start")
-			$status["running"] = "true";
+		if (isset($_REQUEST["a"]))
+		{
+			if ($_REQUEST["a"] == "start")
+				$status["running"] = "true";
 	
-		if ($_REQUEST["a"] == "stop")
-			$status["running"] = "false";
+			if ($_REQUEST["a"] == "stop")
+				$status["running"] = "false";
 
-		if ($_REQUEST["a"] == "update")
-			die(header("location: ?p=camera"));
+			if ($_REQUEST["a"] == "update")
+				die(header("location: ?p=camera"));
+		}	
 	}
 
 	if ($currentPage == "files")
 	{
-		if ($_REQUEST["a"] == "delete")
-		{
-			system("rm ".$status["storagePath"]."/".$_REQUEST["file"]);
-		}
+		if (isset($_REQUEST["a"]))
+			if ($_REQUEST["a"] == "delete")
+			{
+				system("rm ".$status["storagePath"]."/".$_REQUEST["file"]);
+			}
 		$dir = opendir($path=($status["storagePath"]));
 		while ($dat = readdir($dir))
 			if (is_file($path."/".$dat))
